fix: resolve all clippy warnings and improve code quality
- Remove unnecessary type casts (u16 -> u16, f32 -> f32)
- Fix documentation formatting with proper list item indentation
- Implement Default trait for structs with new() methods:
  * BaseObject, PhotonScene, Scene
  * FalseColourMaterial, PhotonMappingMaterial
- Replace manual Default impl with derive macro where applicable
- Change function parameters from &Vec<T> to &[T]
- Remove unnecessary clone() calls on Copy types
- Remove needless reference operators where compiler dereferences automatically
- Replace manual range checks with idiomatic .contains() method
- Use .nth(n) instead of .skip(n).next() for iterator operations
- Replace .flatten() with .map_while(Result::ok) for better error handling
- Refactor functions with too many arguments:
  * Create IntersectionData struct in PolymeshObject
  * Create QuadraticCoefficients struct for Quadratic equation parameters

+++ src/core/framebuffer.rs
@@ -89,7 +89,7 @@ impl FrameBuffer {
 
         let diff = if max - min == 0.0 { 1.0 } else { max - min };
 
-        let ppm_writer = PPMWriter::new(self.width as u16, self.height as u16);
+        let ppm_writer = PPMWriter::new(self.width, self.height);
         ppm_writer.write_file(filename, &self.framebuffer, |pixel| {
             let red = (((pixel.colour.r - min) / diff) * 255.0) as u8;
             let green = (((pixel.colour.g - min) / diff) * 255.0) as u8;
@@ -110,7 +110,7 @@ impl FrameBuffer {
             });
         let diff = if max - min == 0.0 { 1.0 } else { max - min };
 
-        let ppm_writer = PPMWriter::new(self.width as u16, self.height as u16);
+        let ppm_writer = PPMWriter::new(self.width, self.height);
         ppm_writer.write_file(filename, &self.framebuffer, |pixel| {
             let depth = (((pixel.depth - min) / diff) * 255.0) as u8;
             (depth, depth, depth)
